Allow for iterative processing
Don't get rid of original transparency before floodfilling. This
allows us to incrementally add regions to become transparent. However,
since floodfill doesn't work with transparent colors properly, I have
to handle it specially. (For now, I'm guessing that if the color is of
the form (r,g,b,0), I can just turn it to (r,g,b,1) and be hunkydory).